reenviar convite

+++ routes/web.php
@@ -18,9 +18,6 @@ Route::get('/', function () {
 });
 
 
-Route::get('/test', 'Admin\\DocumentoController@test')->name('test');
-
-
 Auth::routes();
 
 Route::get('/admin/convites', 'Admin\\ConviteController@index')->name('convites');
@@ -32,23 +29,25 @@ Route::get('/admin/unidades/{id}/edit', 'Admin\\UnidadeController@edit')->name('
 
 Route::get('/admin/tiposdocumento', 'Admin\\TipoDocumentoController@index')->name('tiposdocumento');
 Route::get('/admin/assuntos', 'Admin\\AssuntoController@index')->name('Assuntos');
-Route::get('/documentos', 'Admin\\DocumentoController@index')->name('documentos');
+Route::get('/admin/documentos', 'Admin\\DocumentoController@index')->name('documentos');
 
-Route::get('/documentos/publicar', 'Admin\\DocumentoController@create')->name('publicar');
-Route::post('/documentos/publicar', 'Admin\\DocumentoController@store')->name('enviar');
+Route::get('/admin/documentos/publicar', 'Admin\\DocumentoController@create')->name('publicar');
+Route::post('/admin/documentos/publicar', 'Admin\\DocumentoController@store')->name('enviar');
 
-Route::get('/documento/{id}', 'Admin\\DocumentoController@show')->name('documento');
-Route::delete('/documentos/{id}', 'Admin\\DocumentoController@destroy')->name('delete');
+Route::get('/admin/documento/{id}', 'Admin\\DocumentoController@show')->name('documento');
+Route::delete('/admin/documentos/{id}', 'Admin\\DocumentoController@destroy')->name('delete');
 
 
 Route::get('/admin/usuarios/{id}/editar', 'Admin\\UsuarioController@edit')->name('usuario-edit');
 Route::get('/admin/usuarios', 'Admin\\UsuarioController@index')->name('usuarios');
 Route::get('/admin/usuarios/convidar', 'Admin\\UsuarioController@convidar')->name('usuario-convidar');
+Route::get('/admin/usuarios/reenviar-convite/{id}', 'Admin\\UsuarioController@reenviarConvite')->name('usuario-reconvidar');
 
-Route::post('admin/usuarios', 'Admin\\UsuarioController@store')->name('usuario-store');
-Route::post('admin/usuarios/create', 'Admin\\UsuarioController@create')->name('usuario-create');
-Route::get('admin/usuarios/pesquisar', 'Admin\\UsuarioController@search')->name('usuario-search');
-Route::post('admin/usuarios/pesquisar', 'Admin\\UsuarioController@search')->name('usuario-search');
+Route::post('/admin/usuarios', 'Admin\\UsuarioController@store')->name('usuario-store');
+Route::post('/admin/usuarios/create', 'Admin\\UsuarioController@create')->name('usuario-create');
+Route::get('/admin/usuarios/pesquisar', 'Admin\\UsuarioController@search')->name('usuario-search');
+Route::post('/admin/usuarios/pesquisar', 'Admin\\UsuarioController@search')->name('usuario-search');
+Route::get('/admin/usuarios/delete/{id}', 'Admin\\UsuarioController@destroy')->name('usuario-delete');
 
 Route::get('/home', 'HomeController@index')->name('home');
